requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.1
aiohttp>=3.9.0
pydantic>=2.6.0

# OCR and classification
paddleocr>=2.7.0,<3.0.0
paddlepaddle>=2.6.0,<3.0.0
ultralytics>=8.2.0